src/plot_data_extraction/lanenet/requirements.txt
absl-py==0.5.0
astor==0.7.1
backcall==0.1.0
bleach==1.5.0
certifi==2018.8.24
cffi==1.11.5
cycler==0.10.0
decorator==4.3.0
easydict==1.6
gast==0.2.0
glog==0.3.1
grpcio==1.15.0
html5lib==0.9999999
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Markdown==3.0.1
matplotlib==2.0.2
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.2
olefile==0.46
opencv-python==3.2.0.7
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-gflags==3.1.2
pytz==2018.5
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
torch==0.4.1.post2
torchvision==0.2.1
tqdm==4.26.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
